requirements.txt
pandas
numpy
psycopg2
sqlalchemy
request
tabpy-server
### System libraries ###
#time
#json
#sys
#csv